(ADSR)
- impl ADSR via VCA (code from Jason). Need to test out in Rx demo.
